fix: extension 분리 및 asset name 수정
