+++ koinOwner/Resources/Images.xcassets/icn_main_menu.imageset/Contents.json
