+++ koinOwner/Resources/Images.xcassets/icn_recommended_menu.imageset/Contents.json
