+++ koinOwner/Resources/Images.xcassets/icn_set_menu.imageset/Contents.json
