+++ koinOwner/Resources/Images.xcassets/icn_side_menu.imageset/Contents.json
