+++ koinOwner/Sources/Common/Components/CustomNavigationBar.swift
@@ -140,18 +140,6 @@ struct BackButton: View {
     }
 }
 
-// Swipe to dismiss 기능 활성화.
-extension UINavigationController: UIGestureRecognizerDelegate {
-    open override func viewDidLoad() {
-        super.viewDidLoad()
-        interactivePopGestureRecognizer?.delegate = self
-    }
-
-    public func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
-        return viewControllers.count > 1
-    }
-}
-
 #Preview {
     MainNavigationBar()
 }
